requirements.txt
streamlit
PyPDF2
SQLAlchemy
google-generativeai
bcrypt
python-dotenv